+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Version 0.28.0 (unreleased)
 
+## Fixes
+
+- Free Lossless Audio Codec (FLAC) media type
+- Matroska 3D Video (MK3D) media type
+- Matroska Audio (MKA) media type
+- Matroska Video (MKV) media type
+
 # Version 0.27.0 (2025-05-09)
 
 ## New formats support